Pass empty options to input if none provided
Previous to this commit, a call to OpenInput() would pass a nil value to
for the options parameter in a subsequent call to OpenInputWithOptions().
This caused a invalid reference. This commit builds an empty options
Pair that is passed to OpenInputWithOptions().

This commit also changes the format_go112.go OpenInputWithOptions
function to use a C.char type for the cfilename